Unify split-path distinctness check across layers
The connection-config schema rejected an inbound_path/outbound_path pair
only when the two strings were byte-identical, while open() rejected any
pair resolving to one directory after textual normalization. A config
like /x and /x/ therefore parsed as a valid split and failed only later
at connect time.

Extract the comparison-only normalizer pathsResolveToSameDir into a
shared util and apply it in the config differ refine, so the schema and
the live connection give the same verdict; open() keeps its own check to
also guard a caller that constructs a connection directly. Invert the
test that pinned the byte-identical accept and parametrize the
now-rejected collision forms; update the file-sync spec and exchange
reference to describe the single shared rule.
