+++ docs/EXCHANGE_REFERENCE.md
@@ -340,7 +340,7 @@ A *split* configuration for deployments where the inbound and outbound folders a
 Rules:
 
 - Set both halves together, or neither; `inbound_path` and `outbound_path` are mutually exclusive with the single `path` (filedrop) / `server.path` (sftp).
-- The two directories must differ. This is checked byte-for-byte at config validation and again, after textual normalization, when the connection opens (both channels): differences that are only redundant separators, `.` segments, a trailing slash, or -- for filedrop -- backslashes are caught. Equivalences that cannot be settled client-side are NOT caught and are the operator's responsibility to keep distinct: `..` segments (which may cross a symlink), case (Windows filesystems are case-insensitive), and -- for SFTP -- a relative path versus the absolute path it expands to under the login home.
+- The two directories must differ. This is checked by the same textual normalization at config validation and again when the connection opens (both channels), so a pair that resolves to one directory fails to parse rather than only failing later at connect time: differences that are only redundant separators, `.` segments, a trailing slash, or -- for filedrop -- backslashes are caught. Equivalences that cannot be settled client-side are NOT caught and are the operator's responsibility to keep distinct: `..` segments (which may cross a symlink), case (Windows filesystems are case-insensitive), and -- for SFTP -- a relative path versus the absolute path it expands to under the login home.
 - A split configuration requires retain mode (`options.retain_files: true`), which in turn requires `lockless_rendezvous: true` and `timestamp_in_filename: true`. Retain mode is what keeps the exchange working across a one-way or delete-suppressing bridge and preserves an auditable record of what flowed each way; see [docs/spec/FILE_SYNC.md](spec/FILE_SYNC.md#split-inboundoutbound-directories).
 - filedrop paths must be absolute (as `path` is). SFTP paths may be absolute or relative, as `server.path` allows. Retain mode requires a fresh directory, which is enforced for *both* directories; an SFTP login-home is usually not empty, so pointing a split directory at it will often be rejected by that check.
