+++ docs/spec/FILE_SYNC.md
@@ -72,7 +72,7 @@ By default one directory is both read and written. A *split* configuration names
 
 **Split mode requires retain mode** -- a deliberate constraint, not a mechanical necessity in every topology. On a shared store a delete-mode sender could poll its outbound (which is the peer's inbound) and observe the peer's delete; but that only holds when deletions propagate back, which fails across the one-way or delete-suppressing bridge that motivates separate folders. Retain mode signals with ack/abort markers that are always carried forward, and keeping the files directly serves the separate-folders goal of an auditable record of what flowed each way. Lock rendezvous is excluded for the same reason it is excluded from retain mode generally: its lock/tiebreaker file must exist in one shared location for the delete-based handshake. Because retain mode already implies lockless rendezvous and timestamped filenames, **only the lockless+retain code paths ever observe split directories** -- the lock and delete-mode paths are unreachable with an outbound directory configured (rejected at config validation), so they stay on `this.path` unchanged.
 
-**Both directories must be fresh.** Retain mode's [fresh-directory precondition](#preconditions-for-a-correct-exchange) is enforced on the inbound *and* the outbound directory at `synchronize()` entry: peer files never land in outbound (the peer writes to its own outbound, which is this party's inbound), so any protocol-grammar file there is this party's own leftover from a crashed prior session and fails the clean-start guard exactly as on the inbound side. An orphaned `temp-*.tmp` is swept from both directories; a foreign file in either is snapshotted and tolerated. The two directories must differ, checked at config validation (byte-identical) and again at `open()` after textual normalization on both channels (`pathsResolveToSameDir`: redundant separators, `.` segments, trailing slashes, and -- filedrop -- backslashes, so e.g. `/x` vs `/x/`, `/x//y` vs `/x/y`, and `/x/./y` vs `/x/y` are all caught). The normalization is textual only: `..` segments (which may cross a symlink), case (Windows is case-insensitive), and a relative-vs-absolute SFTP path under the unknown login home cannot be resolved client-side and remain the operator's responsibility, not a guaranteed reject.
+**Both directories must be fresh.** Retain mode's [fresh-directory precondition](#preconditions-for-a-correct-exchange) is enforced on the inbound *and* the outbound directory at `synchronize()` entry: peer files never land in outbound (the peer writes to its own outbound, which is this party's inbound), so any protocol-grammar file there is this party's own leftover from a crashed prior session and fails the clean-start guard exactly as on the inbound side. An orphaned `temp-*.tmp` is swept from both directories; a foreign file in either is snapshotted and tolerated. The two directories must differ, checked by one shared textual-normalization rule (`pathsResolveToSameDir`) at both config validation and `open()`, so the schema and the live connection give the same verdict (the `open()` check additionally guards a caller that constructs a connection directly, bypassing the schema). On both channels it catches redundant separators, `.` segments, trailing slashes, and -- filedrop -- backslashes, so e.g. `/x` vs `/x/`, `/x//y` vs `/x/y`, and `/x/./y` vs `/x/y` are all caught. The normalization is textual only: `..` segments (which may cross a symlink), case (Windows is case-insensitive), and a relative-vs-absolute SFTP path under the unknown login home cannot be resolved client-side and remain the operator's responsibility, not a guaranteed reject.
 
 ## Phases and legal directory contents
